fix: move constraint checking to sql/query package
