+++ sql/query/create_test.go
@@ -32,6 +32,16 @@ func TestCreateTable(t *testing.T) {
 		{"With primary key", "CREATE TABLE test(foo TEXT PRIMARY KEY)", false},
 		{"With field constraints", "CREATE TABLE test(foo.a[1][2] TEXT primary key, bar[4][0].bat INTEGER not null, baz not null)", false},
 		{"With no constraints", "CREATE TABLE test(a, b)", false},
+		{"With coherent constraint(common)", "CREATE TABLE foo(a DOCUMENT, a.b ARRAY, a.b[0] TEXT);", false},
+		{"With coherent constraint(any)", "CREATE TABLE foo(a, a.b[0] TEXT);", false},
+		{"With coherent constraint(document)", "CREATE TABLE foo(a DOCUMENT, a.b TEXT);", false},
+		{"With coherent constraint(array)", "CREATE TABLE foo(a ARRAY, a[0] TEXT);", false},
+		{"With incoherent constraint(common)", "CREATE TABLE foo(a INTEGER, a.b[0] TEXT);", true},
+		{"With incoherent constraint(common)", "CREATE TABLE foo(a DOCUMENT, a.b[0] TEXT, a.b.c TEXT);", true},
+		{"With incoherent constraint(common)", "CREATE TABLE foo(a DOCUMENT, a.b.c TEXT, a.b[0] TEXT);", true},
+		{"With incoherent constraint(document)", "CREATE TABLE foo(a INTEGER, a.b TEXT);", true},
+		{"With incoherent constraint(array)", "CREATE TABLE foo(a INTEGER, a[0] TEXT);", true},
+		{"With duplicate constraints", "CREATE TABLE foo(a INTEGER, a TEXT);", true},
 	}
 
 	for _, test := range tests {
